refactor controller to independent folder
