+++ ClawTwo/src/modules/leetcode/leetcode.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { LeetcodeService } from './leetcode.service';
-import { LeetcodeController } from './leetcode.controller';
+import { LeetcodeController } from '../../controllers/leetcode/leetcode.controller';
 
 @Module({
   imports: [],
